Move multipart parsing out of the client implementation
